Revert "Store seed phrase backup file on iCloud Drive"
