+++ ConcordiumWallet/Features/NewOnboardingFlow/ImportSeedBasedWallet/ImportWalletView.swift
@@ -19,7 +19,6 @@ struct ImportWalletView: View {
     enum Flow {
         case recoverPhraseInput
         case recoverWithWalletKey
-        case recoverwithSeedPhraseBackup
     }
     
     @State private var flow: Flow? = nil
@@ -55,14 +54,7 @@ struct ImportWalletView: View {
                         })),
                         tag: Flow.recoverWithWalletKey,
                         selection: $flow) { EmptyView() }
-
-                    NavigationLink(
-                        destination: ImportSeedPhraseBackupFileView(viewModel: BackupFileViewModel.init(recoveryService: defaultProvider.recoveryPhraseService(), onValidPhrase: { phrase in
-                            self.recoveryPhrase = phrase
-                        })),
-                        tag: Flow.recoverwithSeedPhraseBackup,
-                        selection: $flow) { EmptyView() }
-
+                    
                     optionsView()
                 }
                 
@@ -149,19 +141,6 @@ struct ImportWalletView: View {
             }
             .onTapGesture {
                 self.flow = .recoverWithWalletKey
-            }
-            
-            Divider()
-                .tint(Color("black_secondary"))
-            
-            HStack {
-                Text("Import via iCloud backup")
-                    .font(.satoshi(size: 14, weight: .regular))
-                Spacer()
-                Image(systemName: "arrow.right").tint(Color.Neutral.tint1)
-            }
-            .onTapGesture {
-                self.flow = .recoverwithSeedPhraseBackup
             }
         }
         .frame(maxWidth: .infinity)
