+++ ConcordiumWallet/Features/RecoveryPhrase/RecoveryPhraseService.swift
@@ -8,7 +8,6 @@
 
 import MnemonicSwift
 import Combine
-import Foundation
 
 protocol RecoveryPhraseServiceProtocol {
     func generateRecoveryPhrase() -> Result<RecoveryPhrase, Error>
@@ -22,8 +21,6 @@ protocol RecoveryPhraseServiceProtocol {
     func store(recoveryPhrase: RecoveryPhrase, with pwHash: String) async throws -> Seed
     
     func store(walletPrivateKey: String, with pwHash: String) async throws -> Seed
-    
-    func isValidWalletPrivateKey(key: String) -> Bool
 }
 
 extension RecoveryPhraseServiceProtocol {
@@ -96,49 +93,5 @@ private extension RandomAccessCollection where Element: Equatable {
         }
         
         return result
-    }
-}
-
-extension RecoveryPhraseServiceProtocol {
-    func isValidWalletPrivateKey(key: String) -> Bool {
-        let hexPattern = "^[0-9a-fA-F]{128}$"
-        
-        do {
-            let regex = try NSRegularExpression(pattern: hexPattern)
-            let range = NSRange(location: 0, length: key.utf16.count)
-            
-            if regex.firstMatch(in: key, options: [], range: range) != nil {
-                let decodedBytes = try hexStringToByteArray(key)
-                return decodedBytes.count == 64
-            } else {
-                return false
-            }
-        } catch {
-            return false
-        }
-    }
-
-
-    
-    private func hexStringToByteArray(_ hex: String) throws -> [UInt8] {
-        guard hex.count % 2 == 0 else {
-            throw NSError(domain: "Invalid hex string: length must be even", code: 0, userInfo: nil)
-        }
-
-        var bytes = [UInt8]()
-        var index = hex.startIndex
-        while index < hex.endIndex {
-            let nextIndex = hex.index(index, offsetBy: 2)
-            let byteString = String(hex[index..<nextIndex])
-            
-            if let byte = UInt8(byteString, radix: 16) {
-                bytes.append(byte)
-            } else {
-                throw NSError(domain: "Invalid hex string", code: 0, userInfo: nil)
-            }
-            
-            index = nextIndex
-        }
-        return bytes
     }
 }
