+++ CryptoX.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "875eb59d06ccd21360debbf3d0f5ba3c06e49d82cb68da712cf6ffb9a537a36d",
+  "originHash" : "7a9fce0a35a9182a0c091da1324f118293d6943f738c84804491ce3997f8f6ae",
   "pins" : [
     {
       "identity" : "abseil-cpp-binary",
@@ -467,15 +467,6 @@
       "state" : {
         "revision" : "4aa89e682f8d7ab1515d85d0797f0d25306bb56a",
         "version" : "1.0.5"
-      }
-    },
-    {
-      "identity" : "zipfoundation",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/weichsel/ZIPFoundation",
-      "state" : {
-        "revision" : "02b6abe5f6eef7e3cbd5f247c5cc24e246efcfe0",
-        "version" : "0.9.19"
       }
     }
   ],
