+++ ProdTestNet.entitlements
@@ -8,18 +8,6 @@
 	<array>
 		<string>applinks:cryptox</string>
 	</array>
-	<key>com.apple.developer.icloud-container-identifiers</key>
-	<array>
-		<string>iCloud.com.pioneeringtechventures.cryptox</string>
-	</array>
-	<key>com.apple.developer.icloud-services</key>
-	<array>
-		<string>CloudDocuments</string>
-	</array>
-	<key>com.apple.developer.ubiquity-container-identifiers</key>
-	<array>
-		<string>iCloud.com.pioneeringtechventures.cryptox</string>
-	</array>
 	<key>com.apple.security.application-groups</key>
 	<array>
 		<string>group.reown.testnet</string>
